requirements.txt
sdx-common>=0.5.1
structlog==17.2.0
cryptography==1.9
PyJWT==1.5.2
pika==0.10.0